Also refactored the landing page
